Intermediate Step to throw out reversion

+++ apps/landmatrix/synchronization/investor.py
@@ -1,12 +1,12 @@
 from typing import Union
 
-import reversion
-
-from apps.landmatrix.models import Investor, InvestorVentureInvolvement
 from apps.landmatrix.models import (
     HistoricalInvestorVentureInvolvement,
     HistoricalInvestor,
 )
+from apps.landmatrix.models import Investor, InvestorVentureInvolvement
+from apps.landmatrix.models.versions import Revision, Version
+from apps.landmatrix.synchronization.helpers import calculate_new_stati
 
 CLASSIFICATIONS_MAP = {
     "10": "PRIVATE_COMPANY",
@@ -49,32 +49,53 @@ def histvestor_to_investor(histvestor: Union[HistoricalInvestor, int]):
     investor.homepage = histvestor.homepage or ""
     investor.opencorporates = histvestor.opencorporates_link or ""
     investor.comment = histvestor.comment or ""
-
-    status = histvestor.fk_status_id
     investor.modified_at = histvestor.history_date
     investor.old_id = histvestor.id
 
+    status = histvestor.fk_status_id
+
     # check involvements
-    involvs = HistoricalInvestorVentureInvolvement.objects.filter(fk_venture=histvestor)
-    for involve in involvs:
+    for involve in HistoricalInvestorVentureInvolvement.objects.filter(
+        fk_venture=histvestor
+    ):
         try:
             Investor.objects.get(id=involve.fk_investor.investor_identifier)
         except Investor.DoesNotExist:
             histvestor_to_investor(involve.fk_investor)
 
-    with reversion.create_revision():
-        _create_involvements_for_investor(investor, histvestor)
+    rev1 = Revision.objects.create(
+        date_created=histvestor.history_date,
+        user=histvestor.history_user,
+        comment=histvestor.action_comment or "",
+    )
 
-        investor.save_revision(
-            status,
-            histvestor.history_date,
-            histvestor.history_user,
-            histvestor.action_comment or "",
+    do_save = not investor or investor.status == 1 or status in [2, 3, 4]
+
+    investor.new_status, investor.new_draft_status = calculate_new_stati(
+        investor, status
+    )
+
+    if do_save:
+        # save the actual model
+        # if: there is not a current_model
+        # or: there is a current model but it's a draft
+        # or: the new status is Live, Updated or Deleted
+        investor.save(custom_modification_date=histvestor.history_date)
+
+    Version.create_from_obj(investor, rev1)
+
+    if not do_save:
+        # otherwise update the draft_status of the current_model
+        Investor.objects.filter(pk=investor.pk).update(
+            draft_status=investor.new_draft_status
         )
 
+    _create_involvements_for_investor(investor, histvestor, status, rev1)
 
-def _create_involvements_for_investor(investor, histvestor):
-    InvestorVentureInvolvement.objects.filter(venture=investor).delete()
+
+def _create_involvements_for_investor(investor, histvestor, do_save, revision):
+    if do_save:
+        InvestorVentureInvolvement.objects.filter(venture=investor).delete()
 
     involves = HistoricalInvestorVentureInvolvement.objects.filter(
         fk_venture=histvestor
@@ -86,13 +107,11 @@ def _create_involvements_for_investor(investor, histvestor):
             if histvolvement.investment_type
             else None
         )
-
-        InvestorVentureInvolvement.objects.create(
+        ivi = InvestorVentureInvolvement(
             investor_id=histvolvement.fk_investor.investor_identifier,
             venture_id=investor.id,
             role=ROLE_MAP[histvolvement.role],
             investment_type=types,
-            status=histvolvement.fk_status_id,
             percentage=histvolvement.percentage,
             loans_amount=histvolvement.loans_amount,
             loans_currency_id=histvolvement.loans_currency_id,
@@ -101,6 +120,9 @@ def _create_involvements_for_investor(investor, histvestor):
             comment=histvolvement.comment or "",
             old_id=histvolvement.pk,
         )
+        if do_save:
+            ivi.save()
+        Version.create_from_obj(ivi, revision)
 
 
 ROLE_MAP = {"ST": "PARENT", "IN": "LENDER"}
